Review feedback + reorganize components
Sorry for the huge diff - I just felt it necessary to reorganize these components to help me better understand what's going on.

Summary of changes in this commit:

- Remove the ExternalOverview component - it's not being used anywhere.
- Move CSS styles into SASS files.
- Rename `desktopView` to `desktop-view.component` - helps make things consistent and also has the benefit of allowing i18next to automatically extract translations from the component.
- Extract the `CommonDataTable` component into its own file.
- Remove `isPatientSummaryDashboard` prop - I can use the insertSeparator prop to toggle styles based on whether the CommonDataTable is used in the Recent Results overview or in the Test Results overview.
