+++ packages/esm-patient-test-results-app/src/desktop-view/desktop-view.scss
@@ -1,0 +1,19 @@
+.grid {
+  display: grid;
+  grid-template-columns: 600px 1fr;
+  gap: 20px;
+  height: calc(100vh - 48px);
+  width: 100%;
+}
+
+.overflow-border {
+  height: 100%;
+  width: 100%;
+  position: relative;
+  overflow: auto;
+}
+
+.overview {
+  display: grid;
+  gap: 1.5rem;
+}
