+++ packages/esm-patient-test-results-app/src/desktop-view/index.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { Switch, Route, BrowserRouter } from 'react-router-dom';
 import { testResultsBasePath } from '../helpers';
-import DesktopView from './desktopView';
+import DesktopView from './desktop-view.component';
 
 const RoutedDesktopView = ({ match, ...props }) => {
   return <DesktopView {...props} {...match.params} />;
@@ -17,4 +17,5 @@ const DashboardRoot = ({ basePath, patientUuid }) => (
     </Switch>
   </BrowserRouter>
 );
+
 export default DashboardRoot;
