+++ packages/esm-patient-test-results-app/src/overview/common-overview.scss
@@ -1,5 +1,13 @@
 @import "~@openmrs/esm-styleguide/src/vars";
 
+.card {
+  border: 1px solid $ui-03;
+}
+
+.activeCard {
+  border: 2px solid $brand-teal-01;
+}
+
 .cardHeader {
   &:after {
     content: "";
@@ -10,14 +18,8 @@
   }
 }
 
-.activeCard {
-  border: 1px solid $brand-teal-01;
-}
-
-.tableContainer :global(.bx--data-table-header) {
-  padding: 0.75rem 1rem 0.25rem 1rem;
-
-  *:not(:first-child) {
-    margin: 0.25rem 0rem;
+.columns {
+  col {
+    width: 33.3%;
   }
 }
