--- packages/esm-patient-test-results-app/src/overview/external-overview.component.scss
@@ -1,7 +1,0 @@
-@import "~@openmrs/esm-styleguide/src/vars";
-@import "~carbon-components/src/globals/scss/vars";
-@import "~carbon-components/src/globals/scss/mixins";
-
-.widgetCard {
-  border: 1px solid $ui-03;
-}
