+++ packages/esm-patient-test-results-app/src/overview/helpers.tsx
@@ -4,7 +4,6 @@ import dayjs from 'dayjs';
 import styles from './lab-results.scss';
 import { TableRow } from 'carbon-components-react';
 import { OBSERVATION_INTERPRETATION } from '@openmrs/esm-patient-common-lib';
-import { useLayoutType } from '@openmrs/esm-framework';
 
 export function formatDate(date: Date) {
   return dayjs(date).format('DD - MMM - YYYY · HH:mm');
@@ -20,20 +19,6 @@ export const Main = ({ className = '', ...props }) => <main {...props} className
 
 export const RecentResultsGrid = (props) => {
   return <div {...props} className={styles['recent-results-grid']} />;
-};
-
-export const Card = ({ ...props }) => {
-  const { isActiveCard, isPatientSummaryDashboard } = props;
-  const isTablet = useLayoutType() === 'tablet';
-
-  return (
-    <div
-      {...props}
-      className={`${styles.card} ${isPatientSummaryDashboard ? '' : styles['outlined-card']} ${
-        isActiveCard && !isTablet && !isPatientSummaryDashboard ? styles['active-card'] : ''
-      }`}
-    />
-  );
 };
 
 export const Separator = ({ ...props }) => <div {...props} className={styles.separator} />;
