+++ packages/esm-patient-test-results-app/translations/en.json
@@ -5,8 +5,10 @@
   "moreResultsAvailable": "More results available",
   "recentResults": "Recent Results",
   "recentTestResults": "recent test results",
-  "seeAllResults": "See all results",
+  "testResults": "test results",
+  "timeline": "Timeline",
   "timeOfTest": "Time of Test",
+  "trend": "Trend",
   "trendlineRangeSelector1Day": "1 day",
   "trendlineRangeSelector1Month": "1 month",
   "trendlineRangeSelector1Year": "1 year",
